Create the base configuration file in a temporary file and rename it into place so a crash can't result in a corrupted base configuration file.
Always create a base configuration file, and always include the version
number (who knows, the initial version might be useful some day).

By always creating a base configuration file, we can detect a crash
between creating the "WiredTiger" file and creating the base
configuration file, and the create the base configuration file when
we're restarted.

That's a (hopefully minor) API change: an application that initially
created the database without config_base set, or simply without any
wiredtiger_open configuration, so that no base configuration file was
written, which then ran this version of the code with config_base set
or with a wiredtiger_open configuration, would suddenly create a base
configuration file where there was none before. I find that unlikely
enough I'm not going to worry about it.

Remove the check for the base-configuration file already existing when
the database is created. A previous revision in this file (#1775) syncs
the "WiredTiger" file to disk before we create the base configuration
file, so the test be safe, but we're no longer relying on ordering
between the creation of the "WiredTiger" file and the base configuration
file, so the test no longer makes sense.

Add simple helper functions to open/close FILE handles so there's less
memory allocation and errno handling inside more complicated routines.

Reference #1775, #1776, SERVER-17571.