Fix a deadlock in LSM with schema operations.
There is special code in LSM to co-ordinate schema operations on
tables (drop, rename, etc). The code does a dance dropping and
acquiring locks, to allow utility operations to drain for the tree
while waiting for it to close.

We were doing the dance with the schema and dhandle list locks. We
needed to include the table lock, or parallel cursor opens could block:

The cursor open is waiting for the table lock:
__wt_spin_lock src/include/mutex.i:175
__schema_add_table src/schema/schema_list.c:26
__wt_schema_get_table src/schema/schema_list.c:98
__wt_curtable_open src/third_party/wiredtiger/src/cursor/cur_table.c:875
__wt_open_cursor src/session/session_api.c:240

The LSM table drop is waiting for the schema lock:
__wt_spin_lock src/include/mutex.i:175
__lsm_tree_close src/lsm/lsm_tree.c:107
__wt_lsm_tree_drop src/lsm/lsm_tree.c:943
__wt_schema_drop src/schema/schema_drop.c:174
__drop_table src/schema/schema_drop.c:124
__wt_schema_drop src/schema/schema_drop.c:176
__session_drop src/session/session_api.c:528