Fix a deadlock related to handle locking.
If one thread does a __wt_session_release_btree of a handle at the same time
as another thread does a __wt_session_get_btree both wanting exclusive access
to the file. It was possible for one thread to get the dhandle list lock and
wait on the handle lock, which another thread held the handle lock waiting for
the handle list lock.

Temporarily fix by doing a try-lock on the __wt_session_get_btree path, long term
the solution is to get rid of the __conn_dhandle_open_lock method, and split get
and lock into two passes.

Refs BF-716