SERVER-18304 Only call _appendHelper() if update/remove succeeds.
Otherwise the response will contain the "value" field multiple times
because the command result is mutated even when the operation does
not succeed (due to a WriteConflictException). Depending on the client
logic, e.g. using the first occurrence, it may end up claiming to have
deleted a document that was actually deleted by another client.