Add support for bulk load in LSM trees.
This allows us to load into a single btree, using btree bulk load
semantics (single threaded, in order, no logging). Once the load completes
we switch the chunk out for the LSM tree.

It's possible that we could avoid some of the switch logic when closing a
bulk load cursor - since the file is flushed when closing the btree handle.
It's simpler to use the switch logic to update the state of the tree.

Refs WT-1922 SERVER-18321