If the cache is large enough but there are too many dirty pages, don't evict the pages, clean them and leave them in the cache; ref WT-1350.
Abstract out the test for when the cache requires eviction or to reduce
the number of dirty pages into a new function, __wt_cache_status().
Change the page eviction function, __wt_evict() to call that function
after reconciling each page, and leave the page in the cache if the
cache is currently large enough.