Add an eviction_dirty_trigger configuration setting.
Allows applications to maintain a low proportion of dirty pages in
the cache, while allowing clean pages to utilize the cache.

While here, alter the minimum allowed dirty trigger and target to
be set as low as 5 percent. Which may be desired for huge caches.

refs SERVER-16737, WT-1350

e Please enter the commit message for your changes. Lines starting