Merge __wt_eviction_check, __wt_cache_full_check and __wt_cache_wait into __cache_check (a simple cache-size helper function), and a worker function, __wt_cache_full_check.
Add a new parameter to __wt_cache_full_check, set if the caller is
"busy". It's set for callers stalling for some reason (for example,
waiting on a disk read). If the caller is busy, we only use it for a
single page eviction before returning it to its previous task.

Change the page-in function to check on the cache when a thread is
about to stall, and evict a page instead of sleeping.

The way this used to work was a thread would start evicting pages if the
cache was > 95% full or there were eviction target values we wanted to
reach, but would evict a single page and quit evicting as soon as the
cache was < 100% full, regardless of eviction target values.  Now we
continue working until we both reach < 95% and the eviction target
values, that is, we'll throttle threads unless they're currently "busy".

Reference WT-1744.