WT-1969: the recent changes to avoid deepening the entire tree when doing append-style workloads (573bf08) doesn't play well with the code that prevents continual deepening of the tree when splits aren't reducing the memory footprint of the page (967a1df). The latter code sets a flag to prevent further deepening splits of the tree at any page after a deepening split fails to reduce the page's memory footprint by 50%.
Specifically, in the Amazon-#73 workload, once the root page grows large
enough that only a relatively small part of the root page's child pages
are being deepened by the split, the memory footprint isn't sufficiently
reduced for the test, and the root page is not further deepened.  This
leads to a root page with 8M entries, and slow cursor traversal.

Remove the code that prevents further deepening of the tree based on
the reduction in the page's memory footprint: that code was added when
we didn't trust the tree's memory statistics (and we were rushing to
get the 3.0 release out), and those problems seem to have been resolved.