SERVER-19127 Fix race condition in unittest log line capturing.
The approach of this fix is to first make the appender for capturing log lines
thread safe, and then to make it very unlikely that bad things happen if a test
races between logging and detaching the special capture appender.  We make it
unlikely by taking advantage of the fact that appenders are removed by setting a
pointer in a vector to NULL, and make sure that the capture appender is not
destroyed when it is detached, but rather when the test fixture destructor runs.
This means that if logging only occurs while a test fixture exists (likely if
tests don't start threads outside of fixtures), tests will not try to use a
capture appender that has gone out of scope.