WT-1977: To defer the equality comparison until outside the binary search loop, you have to reduce the search regardless of the comparison's return, otherwise you'll move past the match.
Since we're comparing integers and the comparison is cheap, I'm adding an
equality comparison inside the loop.  This makes it harder to generalize
this macro (so we may want to go back to deferring the equality comparison
until outside the loop, which only requires a single test of less-than or
greater-than, but as noted, you have to reduce the search regardless of the
test result, which means a different kind of wrapping loop).

One other fix, test for n == 0 before setting limit to (n - 1) and testing
it against 0.