Add a new flag, WT_SESSION_EVICTION_WORKER, set for eviction-worker sessions, this allows us to decide the sessions that get their own lookaside session cursor. Change the eviction-worker session create code to permit data-handles and to set this flag.
Change the lookaside file code to create lookaside file cursors for
eviction-worker threads and for the handle sweep thread.

More reconciliation deadlocks... Turn eviction off for the session as
soon as it gets near a lookaside file cursor, and don't turn eviction
back on for the session until it's done with (or discarded) the cursor.

Set the no-eviction flag on the sweep thread's session, I don't think
we ever want to tap the sweep thread for eviction.

Have __wt_page_release check the session's no-eviction flag explicitly:
the path where we get into trouble is when we're swapping pages during
a walk, we've got the page we want and we're discarding the page we had
before.