We weren't checking for an error return from __wt_page_release in __evict_walk_file, return an error if __wt_page_release() fails.
Move the update of the slot to immediately after the loop instead of
waiting, we update it regardless of error, so might as well do it
near the loop that's incrementing evict.

Remove the comment about locking "interrupting" the walk from the test
for WT_NOTFOUND, I don't think that can happen.