I agree with Don, there's a bug in the __wt_try_readlock() function where the readers value can overflow into the users value, and so it needs to be masked so it doesn't overflow. Use a masking operation rather than a separate variable.
Rename the WT_RWLOCK.pad field to be "overflow", it's not a padding byte.

Now that I've loaded this all into my head again, take Don's Wiki page
as a starting point and comment the lock structure and how all of this
stuff works, the web page we reference might not be around the next time
we need to figure out what's going on.