SERVER-19212 Hide indexes and collections from stale majority concern reads
Reading from a collection will error out if the snapshot used by the query is
older than the collection creation or the point when the last index was
dropped from a collection.

New indexes do not cause reads to fail, but reads cannot use the new indexes
since they may be incomplete in the snapshot.