Stop trying to drop the lookaside file on database close, for a couple of reasons:
First, I can delete the file, but I can't get the metadata record, it's
never removed.

Second, the natural way to drop the lookaside file is to close the
lookaside session, implicitly closing the cursor, and dropping the
lookaside file in the context of the connection's default session.
(Dropping the lookaside file in the context of the lookaside file's
session works, but doesn't get the metadata record, see above.) Anyway,
if I drop the file in the context of the connection's session, the
lookaside file is added to the connection's session file handle list
(along with the metadata file), and that eventually drops core when
the connection's data handles are closed.

Moving the drop earlier in the database close cycle is messy, because
eviction can be using the lookaside file up to the last minute. During
close, it would be OK to not have a lookaside file, because all updates
will be visible to any evicting/closing thread, but there are going to
be problems in figuring out exactly when we no longer need the lookaside
file.

For now, I'm just not dropping the lookaside file during close.