WT-2045: There are two places in the page-read code where a reading thread calls __wt_cache_eviction_check without first checking the WT_READ_NO_EVICT flag, and that can lead to deadlock. The case I have is a configuration with no eviction-worker threads, lots of eviction pressure, and so the application threads are busy doing eviction.
The eviction server thread splits a page, is getting a hazard pointer
on the parent, and for some reason can't get it immediately. The
eviction server thread then gets pressed into doing more eviction, there
are no pages on the queue because the application threads have emptied
it, and things stall forever while the eviction server looks for pages
on the empty eviction queue.