SERVER-19855 Do not perform shard version checking where not necessary
The code in RangeDeleter and sharding migration and split uses
AutoGetCollectionForRead, which has the side effect of checking for shard
version, based on the context. This causes problems in the cases where we
are transmitting shard version information as part of the context, in
particular during the migration cases.

This change gets rid of these checks and replaces them with plain lock
retrieval.