Fix the WT_SPINLOCK structure declaration for alignment.
Switch to a structure for gcc spinlocks, I couldn't figure out any way
to get alignment for an int. (The structure only has a single field, a
volatile int, so I don't think the code will be any slower).

Remove the typedef on the declaration, it's not needed, the dist scripts
with automatically generate it.