WT-2076 Fix a bug preventing the save/restore eviction path from working. In summary, if we found a row with a set of changes where we had to skip a change, but subsequently found a row with a set of changes where we didn't have to skip a change, but there were not-yet-globally-visible changes, we returned EBUSY and failed reconciliation because the lookaside table flag wasn't set. This led to huge performance slowdowns in some wtperf tests. Reference #2167, #2165.
As part of this fix, configure either the save/restore eviction path or
the lookaside table eviction path, but never both. Specifically, the
save/restore path "writes" pages with the original, on-page key/value
pairs, but the lookaside table path writes (that is, it writes to disk),
pages with the most recently committed change, and we can't switch
between the two paths.

Fix a bug, found by inspection, don't configure the lookaside table for
internal pages, it's leaf-pages only.