SERVER-20385: the original use case of WT_CURSOR.next(random) was to return a point in the tree for splitting the tree, and for that reason, once we found a random page, we always returned the first key on that page in order to make the split easy.
In MongoDB: first, $sample de-duplicates the keys WiredTiger returns,
that is, it ignores keys it's already returned; second, $sample allows
you to set the sample size. If you specify a sample size greater than
the number of leaf pages in the table, the de-duplication code catches
us because we can't return more unique keys than the number of leaf
pages in the table.

Remove the code that returns the first key of the page, always return
as a random a key as we can.