TOOLS-982 remove BSONPath from intents
BSONPath is replaced with different things in differnet contexts

Its original use was as the path to open on disk. Many intents don't
have paths on disk. Paths to files on disk are not represented in
the path member int the realBSONFile and realMetadataFile structs

It was then also use it log and error messages. This use was partially
replaced with the Location member of the intent. For restore, this is
now more fully replaced. dump could probably be brought up to par in its
use of Location.

Additionally it was used to indicate that an intent had either a BSON
part or a Metadata part. This use is not replaced with checking that
the intent contains either a BSONFile or MetadataFile