WT-2207 Check for exclusive access when locking handles.
The conditions when this should succeed are subtle, in particular we have a test that opens a tree for a bulk load, then *in the same session* executes checkpoints.  This seems like a really bad idea, and I am inclined to disallow checkpoints in this situation, but for now, make it work.