Off-by-one: the maximum value for base is one past the last slot, so the test for "are we using the last slot?" is (entries == base), not (entries == base - 1).
Note, this change fixes a test for right-hand descents of the tree as
well. I don't expect that change to make a difference anywhere because
the fast-append path gets it correct, the bug was failing to turn the
descend-right boolean to "off" in some cases where the search wasn't a
descend-right search.