Rework 2a71849, some new verbose messages.
Print out the WT_PAGE address, not the WT_REF, print the address first
so it's easy to sort the output, enhance a couple of the messages with
additional information.

Fix a couple of problems: the __split_parent message was checking if
ref->page was NULL to determine if it's a reverse split, not ref_new,
the verbose message in __split_parent_climb was returning instead of
jumping to the error label (I removed that verbose message, since all
of the functions it calls have their own).