WT-1315.  Support statistics on join cursors.
Statistics for these are a bit different than for a data source or a
connection:

1) The statistics are simple (currently only 3), but range over sets of
indices, so traversing the stats is a two level operation.  Added a private
callback API to advance to the next index when the stats have been
exhausted.

2) cursors are used as single threaded operations, so arrays of stats is to
avoid thread conflicts is not needed.

Minor change: the stats_desc callback changed to return an int, and added
an arg to provide some context.  In the join cursor case, the returned desc
is not a static string, it needs to be created and managed.