WT-2231: Update the WT_REF's page-index hint as we split pages. This can race with a thread setting the hint based on an older page-index, and the change isn't backed out in the case of an error, so there ways for the hint to be wrong; OK because it's just a hint.
I'm doing upates in both __split_parent and __split_ref_move_final, some
of which is duplicated work. The reason is that __split_parent is the
only way to catch reverse splits, and __split_ref_move_final is the best
way to catch created WT_REF structures, and the additional work is minimal.