SERVER-21585 Don't use the lookaside file until the cache is stuck full.
Also only write into the lookaside file from internal (aka eviction)
threads, so application threads don't pay the performance cost.

It is possible that the cache stuck check is too restrictive, but using
only the eviction aggressive flag can result in writing a lot of content
to the lookaside file when pages can still be evicted.