SERVER-21553 Enable fast-path truncate after splits.
A truncate operation attempts to mark leaf pages deleted without reading them
into cache.  One of the conditions that has to be met for that fast-path
truncate of pages is that the leaf page not contain overflow items (or we would
need to read it in order to delete the overflow items).

The "no overflow" flag was not being preseved across internal page splits, so
recent changes to splits were defeating fast-path truncation.

Add tracking of the "no overflow" flag for in-memory page addresses so
fast-path truncates work after internal pages are split.