Some support for WT_READ_SKIP_LEAF has to live in __wt_tree_walk(), it simplifies things to push it all down there.
Split the check for in-memory/on-disk into two parts, depending on what
we access there is at our current point in the code, __wt_ref_is_leaf()
goes back to being a simple on-page cell check.

Pages read for data sampling aren't "useful"; don't update the read
generation of pages already in memory, and if a page is read, set its
generation to a low value so it is evicted quickly.