WT-2231: this is a replay of the changes for the previous WT-2231 branch (see pull request #2325) for the individual commit details. There's an additional bug fix over and above those commits, fixing a clang sanitizer error.
In summary: check the leaf page's parent keys before doing the full
binary search of the leaf page, avoiding the binary search entirely
when the cursor is being re-positioned, at the cost of two additional
searches when the cursor is not being re-positioned. Additionally,
do some work to improve WT_REF page hints.