Instead of creating a fake WT_BLOCK structure, drive the allocation-size argument all the way down the function stack, that way we won't break if the WT_BLOCK structure changes.
Flip the argument order for __wt_block_ckpt_decode() to match the other
functions.

Have __wt_block_ckpt_decode() take a WT_SESSION, rather that doing that
cast outside of the WiredTiger library itself.

Lint: cast szvalue.val in util_list.c to be a size_t, convert to a uint32_t
when we get into the WiredTiger library.