gcc47 with [-Werror=maybe-uninitialized] reports we can end up with type uninitialized in __ref_is_leaf() (based on a call to __wt_ref_info()). It's not really possible because the path where type isn't set is a path where we panic because the WT_ADDR structure has an impossible type.
We already ignore the __wt_ref_info() error return in one path, and
there are only two paths that care about the returned type; remove the
error check from __wt_ref_info() and set type to 0 in the failing case
(the same value we use when there's no WT_REF addr to check), the code
that calls this function already checks addr on return.

This simplifies __ref_is_leaf() slightly, it now returns a boolean
instead of an error code with a boolean pointer argument.