Move __wt_buf_set_printable() from buf.i to bt_misc.c, there's no reason to inline that function, it's an error/verbose print routine.
Change __wt_buf_set_printable() to return a pointer to the (char *)
string created (which is potentially "[Error]"). The change simplifies
some code, we can now make the call to __wt_buf_set_printable() as an
argument to __wt_verbose(), rather than calling it separately and
needing a verbose flag wrapper and/or additional error handling.

Change __wt_buf_set_printable() to take the temporary buffer as the last
argument instead of the second, matches __wt_addr_string().

Change __verify_dsk_row to display the mismatched keys when finding two
keys out-of-order during verify, it's the most common failure I see and
I'm tired of hacking that code in during debug.