WT-1801: Add a directory sync after rollback of a WT_SESSION::rename operation
Change __meta_track_unroll() to flush the from/to directories when
unrolling a rename.

Change __wt_sync_and_rename_fh() and __wt_sync_and_rename_fp() to flush
the directory in which the file exists (don't assume the file is in the
database home directory, although it almost certainly is), as well as to
flush both the from and to directories.