WT-2123: Don't clear allocated memory if not required
The buffers we don't want to waste time clearing are read/write WT_ITEMs.
Change the standard realloc code to take an argument if memory needs to
be cleared. Add a new standard realloc API that clears memory, and
replace __wt_realloc_aligned with a WT_ITEM-specific function that never
clears memory (whether is uses the underlying realloc or posix-memalign
functions).