WT-2361: column-store starting record number error
Don't continue searches from the "current" page if a cihld page we're
waiting on splits.

The problem is a thread holding a hazard pointer on a parent page X, and
waiting on a child page A. Then, A is insert-split, moving a chunk of
A's name-space to a new page B, that follows A in the parent page's
page-index. If the parent page X then does a split into its parent, B
could move to the parent of X, and the thread has to search from the
root of the tree, not just from page X, to find the name-space now in
page B.

I've never seen this bug fire (as far as I know), it's improbable, but
I think it's possible.