WT-2361: column-store starting record number error
If we don't restart in the current page, we have to release it, we're
holding a hazard references. This means all of the jumps to the restart
label release the current page, simplify the world by doing the work at
the lable.

Fix the cursor-random code as well, make it all look the same.