SERVER-22491 relax uptimeEstimate js test
In the original commit, there was a c++ test added, but that test later
was removed when time keeping logic was extracted from the listener.
This backport only removes the bug from the js test (which was the
reason for backporting anyway).

(cherry picked from commit 84b8780894544f1301d92b65075c16da175b4acb)