WT-2397: Cursor traversal from end of the tree skips records.
If we're half-way down the tree, and get a RESTART trying to swap to a
new page, and the reason we get the RESTART is that our page split into
our parent, and caused our parent to split into its parent, we can't use
our parent's page-index, because it's been truncated. Do a full restart.

Simplify the __page_descend_prev() and __page_initial_descent_prev()
functions, they don't need to set the slot, the caller can do that,
and remove an unnecessary local variable.