SERVER-22784: Coverity analysis defect 77722: Unused value
The variable "couple" is assigned, but we jump to a label where it's
assigned again, before it's used. I don't want to change the original
assignment, I think it adds clarity to set couple, couple_orig and ref
in the initial case. Pull the assignment of couple out of the if/else
that holds the jumped-to label, instead.