WT-2107: Add example code including an event handler
My last commit introduced a bug, letting the event handler go out of
scope before a call that used it. Split out the set of the event handler
from the main function, that way we don't have to play games to get the
CUSTOM_EVENT_HANDLER declaration next to the code that initializes it.

Minor #include cleanup while in the area.