WT-2423 Cleanup on error when opening a session handle.
We're not holding locks when sweeping the session's handle list, move
the call of __session_dhandle_sweep() from __session_add_dhandle() to
__session_get_dhandle(), simplifying __session_add_dhandle()'s error
handling as it no longer calls __sesion_discard_dhandle().

Move the fixup of the handle reference count after __session_add_dhandle()
failure to __session_get_dhandle(), that way the increment/decrement of
the reference count are relatively close together and logically within a
few lines of each other.

Remove the final argument of __session_add_dhandle(), it was never used.