SERVER-22831 Queue more leaf pages than internal pages for eviction.
Unless we get aggressive, putting internal pages on the LRU queue is
counterproductive.  That is particularly true for workloads that
transition to read-only, where some tables are not being queried.  In
that case, all of the leaf pages are evicted, and eviction wastes a lot
of effort walking and queuing internal pages that are not actually
evicted.