WT-2345: __wt_evict_file_exclusive_on/off cleanups
When checking if application threads should help with eviction, don't
ignore threads operating in files that can't be evicted at the moment,
only ignore threads that are operating on cache-resident files. (They
might be the same thing at the moment, but they're not the same thing
in general.)

Minor whitespace cleanup.