WT-2391: De-prioritize eviction from indexes
Move the tests for whether or not a page generation number is updated
into __wt_cache_read_gen_bump, it gives us consistent behavior in all
callers of not updating any page generation that's set for forcible
eviction (all callers already checked that), and not updating any page
generation that doesn't need to be updated (eviction didn't check that).

Make the code a bit simpler, change __wt_cache_read_gen_bump to take a
page reference instead of returning a generation number requiring an
assignment.