WT-2391: De-prioritize eviction from indexes
Add the cache's notion of the oldest read generation back in, but
instead of always taking the first slot of the queue, walk the queue
until we find a read generation that's not WT_READGEN_OLDEST.

Rework the 50% cutoff code: First, if at least 50% of the whole queue
is WT_READGEN_OLDEST entries, take all of the WT_READGEN_OLDEST entries,
second, use the calculated oldest generation instead of the first slot
in the queue.

When we've just read a page and are setting the read generation for the
first time, try and put it in the middle of the cache.

When eviction hits a page without a read generation number, restore the
previous behavior of putting it in the middle of the cache.