WT-2391: De-prioritize eviction from indexes
The lowest possible page read-generation has a special meaning, it marks
a page for forcible eviction. If we read in page before doing anything
else, we can end up with a page that looks like it needs forcible
eviction, and salvage isn't prepared to deal with that. It's probably a
bug in salvage that it doesn't handle the page being "released" and then
disappearing, but this gets us most of the way there.