WT-2454: checkpoint_sync=false does not prevent flushes/sync to disk
WiredTiger was doing leaf page flushes even when checkpoint_sync=false,
as well as one other less-interesting checkpoint flush (the checkpoint
of the metadata). Don't flush leaf pages if checkpoint_sync=false, and
check for checkpoint_sync=false in the underlying flush routine instead
of all of the callers -- this isn't worth optimizing out a function call
and it makes things simpler.