WT-2490: search_near() returns wrong key for column-store
In variable-length column-store, updates/deletes are stored in the
insert list. (In short, there needs to be a place to put references for
individual records because we might update/delete a single entry in a
large duplicate set, and the insert list has to exist anyway, so we
used it.) However, if we find a match in the insert list, but it's not
visible to the current transaction for some reason, we need to know if
there's an on-page record with the same record number with a previous
version of the data we can use instead.

We were using an illegal cursor slot value for this purpose, but that
messes up search-near, which needs a valid slot should the entry on
the insert list not be visible.

Add a new flag to the btree cursor, WT_CBT_VAR_ONPAGE_MATCH, that's set
if we find a matching on-page variable length column-store record.

Clarify the column-search routines, specifically set the cursor's
slot and record number for each code path.