WT-2426 Lock all handles for checkpoints in one pass.
In particular, gather all locks (including exclusive locks on old
checkpoints) under the schema / handle list lock so we never reenter
those locks once we have a handle exclusive.

This extends the duration that we hold exclusive locks on checkpoints
to include the write leaves phase.  If that turns out to be a problem,
we could further split out write leaves into a separate pass (that
releases all handles at the end).