WT-2330: in-memory configurations should not create on-disk collection files
Code to remove/create the working directory was checking the directory
existed, before removing it, and using __wt_exist, which has changed.
Remove the test, try and remove the directory if it exists or not.

Rework the code to display the failing command in all cases, rename
a few variables to match other parts of WiredTiger.