WT-2330: in-memory configurations should not create on-disk collection files
In commit 45aa438, I moved the wt utility code that does hot backup
copies from the wt utility into the WiredTiger library, and that's
failing on Windows.

Previously, the wt utility called "open" get a readonly file descriptor
on the object to be copied. When I moved that code into the library, I
changed it to call __wt_open (and eventually the Windows library call
CreateFileA), but still opening readonly.

With that change, the txn04 test started failing:

	test_txn04.test_txn04.test_ops(t1c.remove.row): :
	.\./WiredTigerLog.0000000002: handle-open: CreateFileA: The
	process cannot access the file because it is being used by
	another process.

The problem is the hot backup open of a readonly handle has a different
shared-mode than the WiredTiger existing handle, and that fails.

One workaround/fix is to open the files read/write when doing a copy,
because then the CreateFileA open has the same shared-mode as the
existing handle in WiredTiger. Of course, that breaks if we're ever
copying a file that's currently open readonly in WiredTiger.

The change I'm doing instead is to not pass in the absolute path of the
file to be copied, instead, pass in the name of the file. Then, when we
open it to copy it, if the file is currently open, the backup won't open
a new file handle, so there won't be any collision. If the file isn't
currently open, there obviously won't be any collision.

The failure mode here is if WiredTiger attempts to open a file in
read/write mode, while the readonly backup copy handle is open, then
the WiredTiger thread will see the collision.