WT-2518: LSM does updates on "readonly" file handles
If the chunk is on disk, don't include underlying handles in the checkpoint.
Check the "get handles" function, not the "checkpoint" function, the
get-handles pass happens first.  Checking the "get handles" function is all
we need to do, no further checkpoint calls are done if the handle is not
gathered.