WT-2526: mixing and matching readonly and read/write handles
Change WiredTiger to always open handles in read/write mode unless the
database itself is configured read-only.  The WT_OPEN_READONLY flag is
now OS-specific, and not used above that level.

Requires the creation of a new open flag, WT_OPEN_DIRECTIO: this flag
is set in the OS layer, based on how the connection is configured. It
is also set in the block-manager layer, because that's the only place
we know we're opening a read-only data file (required for support of
the "direct_io=checkpoint" configuration).