WT-2103: add incremental backup testing to format
Add an example of the code to open a cursor for incremental backup.

Wording changes:

Remove the requirement that a checkpoint be run between the full backup
and the incremental backup; it's not really a requirement and it implies
coordination between checkpoints and recovery that could be difficult.

Don't mention the backup database becoming "live" as the terminating
condition for incrementals, once you run recovery, it's over, you cannot
backup to that database copy again.

Note the log file grows rather than "changes in size"; it better not
shrink in size, applications won't be able to handle that.