WT-2548 Cap the amount of data handed to raw compression.
Previously, we had special case code for zlib to bound the compression
ratio to an arbitrary factor.  This is driven by synthetic workloads
where data can be compressed on disk to 1% or less of the in-memory
size.  If WiredTiger allows such on-disk pages to grow, they can end up
larger than the maximum page size permitted in memory and either
immediately force eviction or in extreme cases grow larger than the
total cache size.

This change bounds the amount of data handed to raw compression based on
the maximum page size in memory before attempting in-memory splits.