WT-2223 WT-2553 in_memory documentation and stress testing
Changes to allow running in-memory mode format, automatically; 1 in 20
randomly selected runs are in-memory.

Use 2x the initial load's data size as the minimum cache size. That
appears to always be enough for the column-store initial load, and is
usually enough for the row-store initial load. When not enough, we reset
the number of keys to the number we were able to load, and also set the
insert percentage to a low value and the delete percentage to a higher
value, in order to give ourselves some room in the cache so the run can
reasonably continue.

There is another major change: previous versions of this code used 6-8
times the initial load's data size worth of cache, because we were
creating a relatively large number of huge keys (those are roughly 80K
byte keys). The reason for that was we only generate 1000 random key
lengths (and then select among those lengths), and with a million keys
selecting from 1000 key lengths, we were creating huge keys too often.
This code no longer does that, and both keys and values now select huge
key/value sizes only once in every 2500 key/value pairs.